chore: silence Edge-runtime warnings via instrumentation-node split
`instrumentation.ts` is invoked by Next.js for both Node and Edge
runtimes. Turbopack statically traces the dynamic `await import()`
chain through `@/lib/db/ingestWatcher` → `@/lib/db/ingest` →
migrations / atomicWrite / claudeStatsCache / platform / parser /
costCalculator, all of which use `fs`, `process.cwd()`,
`child_process`, and `process.kill`. That produces a wall of
"Node.js API used in Edge Runtime" warnings on every dev startup —
real visibility-killing noise.

The runtime gate (`if (process.env.NEXT_RUNTIME !== "nodejs") return`)
already prevents anything from actually executing in Edge. The fix
needed is build-time, not runtime: stop Turbopack from compiling
the Node-only graph for Edge in the first place.

Solution: extract the Node logic into `instrumentation-node.ts`
with `'server-only'` at the top. `'server-only'` is Turbopack's
"do not compile for Edge" sigil — it severs the trace at the
dynamic-import boundary. `instrumentation.ts` shrinks to the
runtime gate plus `await import('./instrumentation-node')`.

Verified by running `pnpm run dev`: previous startup printed ~10
"Node.js API used in Edge Runtime" warnings; new startup prints
zero. Worker indexer still spawns, /api/usage still works, all 614
tests pass, typecheck clean.
